Migrated existing solution to atomic design structure
Done to clean up the ever growing folder structure
