+++ .storybook/preview.ts
@@ -1,3 +1,4 @@
+import "../src/style/style.scss"
 import type {Preview} from "@storybook/react";
 
 const preview: Preview = {
