+++ README.md
@@ -1,10 +1,14 @@
 # Windows XP
 
-A simple react application mimicign the Windows XP experience
+A simple react application to mimic the Windows XP experience
 
 # Development
 
 - Developed with ```Node 20.x```
+
+## Project Structure
+
+This app follows the atomic design. This is nicely described [here](https://bradfrost.com/blog/post/atomic-web-design/)
 
 # Future work
