+++ src/main.tsx
@@ -1,10 +1,14 @@
+import "./style/style.scss"
 import {StrictMode} from 'react'
 import {createRoot} from 'react-dom/client'
-import App from './App.tsx'
-import './index.css'
+import {SpeedInsights} from "@vercel/speed-insights/react";
+import {Analytics} from "@vercel/analytics/react";
+import Default from "./components/templates/Default/Default.tsx";
 
 createRoot(document.getElementById('root')!).render(
     <StrictMode>
-        <App/>
-    </StrictMode>,
+        <Default/>
+        <SpeedInsights/>
+        <Analytics/>
+    </StrictMode>
 )
