+++ src/style/01_reset/reset.scss
@@ -1,8 +1,3 @@
-:root {
-    --default-text-color: #f3f9f4;
-    --taskbar-height: 2.5rem;
-}
-
 /* Css reset from https://www.joshwcomeau.com/css/custom-css-reset/ */
 /*
   1. Use a more-intuitive box-sizing model.
@@ -58,12 +53,8 @@ p, h1, h2, h3, h4, h5, h6 {
     isolation: isolate;
 }
 
+// Default body
 body {
-    margin: 0;
     height: 100vh;
     width: 100vw;
-    background-image: url("/windows-xp-background.jpg");
-    background-size: cover;
-    background-position: top;
-    background-repeat: no-repeat;
 }
