+++ src/style/02_settings/settings.scss
@@ -1,0 +1,18 @@
+// Font
+@font-face {
+    
+}
+
+// Global variables
+:root {
+    // ----------------------------------------
+    // FONTS
+
+    // ----------------------------------------
+    // COLORS
+    --default-text-color: #f3f9f4;
+
+    // ----------------------------------------
+    // MEASUREMENTS
+    --taskbar-height: 2.5rem;
+}
