+++ src/style/04_Templates/default.scss
@@ -1,0 +1,6 @@
+.default {
+    background-image: url("../../../public/windows-xp-background.jpg");
+    background-size: cover;
+    background-position: top;
+    background-repeat: no-repeat;
+}
