+++ src/style/05_components/app-shortcut.scss
