+++ src/style/05_components/desktop.scss
