+++ src/style/05_components/error-window.scss
