+++ src/style/05_components/start-button.scss
