+++ src/style/05_components/taskbar.scss
